Added SceneDelegate to control Appflows.
